move tenant allowed method
