+++ tapis-shared-api/src/main/java/edu/utexas/tacc/tapis/sharedapi/utils/TapisRestUtils.java
@@ -2,11 +2,14 @@ package edu.utexas.tacc.tapis.sharedapi.utils;
 
 import org.apache.commons.lang3.StringUtils;
 
+import edu.utexas.tacc.tapis.shared.exceptions.TapisException;
+import edu.utexas.tacc.tapis.shared.exceptions.runtime.TapisRuntimeException;
 import edu.utexas.tacc.tapis.shared.utils.TapisGsonUtils;
 import edu.utexas.tacc.tapis.shared.utils.TapisUtils;
 import edu.utexas.tacc.tapis.sharedapi.dto.ResponseWrapper.RESPONSE_STATUS;
 import edu.utexas.tacc.tapis.sharedapi.responses.RespAbstract;
 import edu.utexas.tacc.tapis.sharedapi.responses.RespBasic;
+import edu.utexas.tacc.tapis.sharedapi.security.TenantManager;
 
 /** This class is a replacement for the RestUtils class in the same package.
  * This class provides an alternative way to generate JSON responses to HTTP
@@ -120,5 +123,33 @@ public class TapisRestUtils
         
         // Correct format.
         return true;
+    }
+    
+    /* ---------------------------------------------------------------------- */
+    /* isAllowedTenant:                                                       */
+    /* ---------------------------------------------------------------------- */
+    /** Determine if the tenant specified in the jwt's tapis/tenant_id claim is
+     * allowed to execute on behalf of the target tenant specified in the 
+     * X-Tapis-Tenant header or the delegation_sub claim.  Neither parameter 
+     * can be null.
+     * 
+     * If the number of allowable tenants becomes too great we may have to 
+     * arrange for a constant time search rather than the current linear search.
+     * 
+     * @param jwtTenantId the tenant assigned in the jwt tapis/tenant_id claim
+     * @param newTenantId the tenant assigned in the X-Tapis-Tenant header
+     * @return true if the jwt tenant can execute on behalf of the new tenant,
+     *         false otherwise
+     * @throws TapisException 
+     * @throws TapisRuntimeException 
+     */
+    public static boolean isAllowedTenant(String jwtTenantId, String newTenantId) 
+     throws TapisRuntimeException, TapisException
+    {
+        // This method will return a non-null tenant or throw an exception.
+        var jwtTenant = TenantManager.getInstance().getTenant(jwtTenantId);
+        var allowableTenantIds = jwtTenant.getAllowableXTenantIds();
+        if (allowableTenantIds == null) return false;
+        return allowableTenantIds.contains(newTenantId);
     }
 }
