+++ tapis-shared-lib/src/main/resources/edu/utexas/tacc/tapis/shared/i18n/TapisMessages.properties
@@ -1814,6 +1814,10 @@ SK_ROLE_NOT_FOUND=SK_ROLE_NOT_FOUND Role {1} not found in tenant {0}.
 SK_GET_ROLE_ID_ERROR=SK_GET_ROLE_ID_ERROR Unable to retrieve the ID for role {0} in tenant {1}: {2}
     # 0 = tenant, 1 = user, 2 = role name, 3 = error message
 SK_CREATE_ASSIGN_ROLE_ERROR=SK_CREATE_ASSIGN_ROLE_ERROR Failed to create role "{2}" and assign it to user "{1}" in tenant {0}: {3}
+    # 0 = jwt tenant, 1 = request tenant, 2 = jwt type
+SK_UNEXPECTED_TENANT_VALUE=SK_UNEXPECTED_TENANT_VALUE The tenant value "{0}" specified the {2} JWT is not compatible with the value "{1}" specified in the request parameters. 
+    # 0 = jwt user, 1 = request user, 2 = jwt type
+SK_UNEXPECTED_USER_VALUE=SK_UNEXPECTED_USER_VALUE The user value "{0}" specified the {2} JWT is not compatible with the value "{1}" specified in the request parameters.
 
     # 0 = tenant id, 1 = error message
 SK_USER_GET_NAMES_ERROR=SK_USER_GET_NAMES_ERROR Unable to retrieve users assigned roles in tenant {0}: {1}
